plot test function using local three-d dependency
- plotting a 2D sinusoidal function to test plotter
- use local three-d crate instead of public one

+++ src/operator_descr.rs
@@ -71,11 +71,11 @@ pub fn default_operator_table() -> OperatorTable {
     let consts = vec![
         ConstantOp {
             symbol: "pi",
-            semantics: std::f64::consts::PI,
+            semantics: std::f64::consts::PI as f32,
         },
         ConstantOp {
             symbol: "e",
-            semantics: std::f64::consts::E,
+            semantics: std::f64::consts::E as f32,
         },
     ];
